Consolidate the duplicated routes behind each Teams fix
Follow-up to the previous commit, which added mechanisms next to ones
that already existed.  Each concern now has exactly one implementation
and the older routes are gone.

Team names.  config.h owns the table; the la.team_short() verb hands it
to Lua, so std.team_announce and upkeep.lua stop spelling "O"/"X" out for
themselves and Teams, Flag and Upkeep all name teams the same way.  The
vestigial kTeamO/kTeamX constants are deleted.

The bar.  bindCooldownVariable (a number that becomes a bar at zero, dead
since it was written) and last commit's bindBarVariable are one binding:
a variable, the trigger value it turns into a bar at, and a live pointer
to the fill duration — a pointer because bindings are built once at game
start while recharge/respawn seconds stay menu-editable.  Nothing reaches
it yet by design: it is the projector object's, and the descriptor-side
plumbing (VarType::BAR, the `bar` monitor field, the Teams respawn bar)
is withdrawn until that lands.  drawBar is the only bar renderer.

Respawn signalling.  The native era had already fixed the dropped-reply
bug with player-to-totem unicasts (MSG_RESPAWN_NOTIFY/MSG_CP_NOTIFY);
the Lua migration went back to replies and carried neither the fix nor
the deletion.  The reply mechanism wins: those constants are retired
beside MSG_FLAG_RETURN, and the radio-level fix stands.  What made the
replies untrustworthy is also gone — nothing auto-answers on a ruleset's
behalf any more.  A handler's return value IS the reply, so a totem that
gives itself to whoever answers hears only players that acted on it.
That needed deliberate, range-gated BONUS/MALUS handlers, which no game
had: pickups were being claimed by the first empty auto-reply.  A test
now fails if a declared pickup slot has no handler.

RSSI.  std.base_respawn's -57 fallback is gone; the gate is the ruleset's
call and both std helpers error without one.  The VM keeps its G_RSSI
guard and gains the matching value operand, so a program can store and
compare readings rather than only test one against a literal — registers
are int16_t to hold both a negative dBm and CP's 0xFF sentinel.  Respawn
stays player-authoritative; no role uses either form yet.

Animation.  One LED runs the strip in 1 s, and one-shots queue (depth 10,
overflow dropped) instead of overwriting: two players respawning together
are shown one after the other, and FLAG's missing-then-taken pair both
play.

Tests: new `strip` suite for the queue, `radio` for the broadcast fan-in,
VM coverage for the signed register (verified failing on uint8_t), and
the games suite checks the std gates and pickup wiring directly.  All
five suites pass.

+++ src/totem/LightAir_TotemVM.h
@@ -70,7 +70,10 @@ private:
 
     // ---- Runtime state ----
     uint8_t  _state = 0;                       // current state index (0-based)
-    uint8_t  _regs[TotemVMDefs::MAX_REGS] = {0};
+    // Signed and 16-bit wide on purpose: a register has to hold both the
+    // 0xFF "neutral owner" sentinel the CP role uses and a negative RSSI
+    // reading, which 8 bits cannot represent distinctly either way round.
+    int16_t  _regs[TotemVMDefs::MAX_REGS] = {0};
     uint16_t _acc = 0;
     uint32_t _timers[TotemVMDefs::MAX_TIMERS] = {0};
